+++ internal/provider/data_source_group.go
@@ -3,56 +3,59 @@ package provider
 import (
 	"context"
 
-	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
-	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 
-	"github.com/slok/terraform-provider-onepasswordorg/internal/provider/attributeutils"
+	"github.com/slok/terraform-provider-onepasswordorg/internal/storage"
 )
 
-type dataSourceGroupType struct{}
+var (
+	_ datasource.DataSource              = &groupDataSource{}
+	_ datasource.DataSourceWithConfigure = &groupDataSource{}
+)
 
-func (d dataSourceGroupType) GetSchema(_ context.Context) (tfsdk.Schema, diag.Diagnostics) {
-	return tfsdk.Schema{
+func NewGroupDataSource() datasource.DataSource {
+	return &groupDataSource{}
+}
+
+type groupDataSource struct {
+	repo storage.Repository
+}
+
+func (d *groupDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_group"
+}
+
+func (d *groupDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
 		Description: `
 Provides information about a 1password group.
 `,
-		Attributes: map[string]tfsdk.Attribute{
-			"name": {
+		Attributes: map[string]schema.Attribute{
+			"name": schema.StringAttribute{
 				Description: "The name of the group.",
 				Required:    true,
-				Validators:  []tfsdk.AttributeValidator{attributeutils.NonEmptyString},
-				Type:        types.StringType,
 			},
-			"description": {
+			"description": schema.StringAttribute{
 				Computed: true,
-				Type:     types.StringType,
 			},
-			"id": {
+			"id": schema.StringAttribute{
 				Computed: true,
-				Type:     types.StringType,
 			},
 		},
-	}, nil
+	}
 }
 
-func (d dataSourceGroupType) NewDataSource(ctx context.Context, p tfsdk.Provider) (tfsdk.DataSource, diag.Diagnostics) {
-	prv := p.(*provider)
-	return dataSourceGroup{
-		p: *prv,
-	}, nil
-}
-
-type dataSourceGroup struct {
-	p provider
-}
-
-func (d dataSourceGroup) Read(ctx context.Context, req tfsdk.ReadDataSourceRequest, resp *tfsdk.ReadDataSourceResponse) {
-	if !d.p.configured {
-		resp.Diagnostics.AddError("Provider not configured", "The provider hasn't been configured before apply.")
+func (d *groupDataSource) Configure(ctx context.Context, req datasource.ConfigureRequest, resp *datasource.ConfigureResponse) {
+	appServices := getAppServicesFromDatasourceRequest(&req)
+	if appServices == nil {
 		return
 	}
 
+	d.repo = appServices.Repository
+}
+
+func (d *groupDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
 	// Retrieve values.
 	var tfGroup Group
 	diags := req.Config.Get(ctx, &tfGroup)
@@ -62,7 +65,7 @@ func (d dataSourceGroup) Read(ctx context.Context, req tfsdk.ReadDataSourceReque
 	}
 
 	// Get group.
-	group, err := d.p.repo.GetGroupByName(ctx, tfGroup.Name.Value)
+	group, err := d.repo.GetGroupByName(ctx, tfGroup.Name.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError("Error getting group", "Could not get group, unexpected error: "+err.Error())
 		return