+++ internal/provider/resource_onepasswordorg_vault_test.go
@@ -5,7 +5,7 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 
 	"github.com/slok/terraform-provider-onepasswordorg/internal/model"
 	"github.com/slok/terraform-provider-onepasswordorg/internal/provider"
@@ -62,7 +62,7 @@ resource "onepasswordorg_vault" "test" {
 	name = ""
 }
 		`,
-			expErr: regexp.MustCompile("Attribute can't be empty"),
+			expErr: regexp.MustCompile("Attribute name string length must be at least 1, got: 0"),
 		},
 	}