+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - Update Go version to v1.23.
+- Update terraform related dependencies.
+- Remove all `terraform-plugin-sdk` usage in favor of the new `terraform-plugin-framework` libs.
 
 ## [v0.5.0] - 2022-07-30
 
@@ -18,7 +20,6 @@
 ### Added
 
 - Vault user access with fine grain permissions.
-
 
 ## [v0.3.0] - 2022-03-17